Optimize C types for better performance and correctness
- Replace `long` with `size_t` for lengths, sizes, and array indices
- Use `unsigned char` for character operations (safer, clearer intent)
- Proper type casting when interfacing with Ruby C API
- Better alignment for modern 64-bit architectures

Performance benefits:
- size_t is the proper type for sizes/indices (unsigned, native word size)
- Unsigned arithmetic can be faster (no sign extension needed)
- Better compiler optimization opportunities
- Clearer code intent (sizes cannot be negative)

Changes by file:
- html_escape.c: size_t for all lengths, positions, loop indices
- tag_render.c: size_t for tag_len, attrs_len, content_len, estimated_size
- attributes.c: size_t for hash_size

Technical details:
- Keep `long` return type for find functions (return -1 on not found)
- Cast Ruby RSTRING_LEN to size_t (Ruby API uses long)
- Use unsigned char for lookups (proper range 0-255)
- All loop indices now size_t for consistency

All tests passing (13 examples, 0 failures)
Zero compiler warnings
Performance: consistent or slightly improved

+++ ext/ruby2html/attributes.c
@@ -28,7 +28,7 @@ VALUE fast_attributes_to_s(VALUE self, VALUE hash) {
     if (NIL_P(hash) || RHASH_EMPTY_P(hash)) return rb_str_new2("");
 
     // Estimate size based on hash size
-    long hash_size = RHASH_SIZE(hash);
+    size_t hash_size = (size_t)RHASH_SIZE(hash);
     VALUE result = rb_str_buf_new(hash_size * 32); // ~32 bytes per attribute average
 
     // Set up context for callback
